ctr: move push and pull to images command
