+++ cmd/ctr/commands/images/pull.go
@@ -1,4 +1,4 @@
-package main
+package images
 
 import (
 	"fmt"
