+++ cmd/ctr/commands/images/push.go
@@ -1,4 +1,4 @@
-package main
+package images
 
 import (
 	gocontext "context"
